:hammer: Abstract common logic into CacheManager interface
